Agregados y correcciones generales
-Se solucionar bugs en agregar al carrito desde la página default.

-Ahora si agregamos más de un libro igual al carrito, estos no se mostraran dos veces, sino que con un contador en el mismo.

-Se puede aumentar la cantidad de los libros que queremos comprar mediante dos botones "-" y "+".

+++ Negocio/AccesoLibrosPorCarrito.cs
@@ -39,19 +39,46 @@ namespace Negocio
 
             return lista;
         }
+        public bool ExisteElLibro(int idCarrito,int IdLibro)
+        {
+            AccesoDatos datos = new AccesoDatos();
+            try
+            {
+                datos.Conectar();
+                datos.Consultar("SELECT COUNT(*) FROM LibrosPorCarrito WHERE IDCarrito = @IDCarrito AND IDLibro = @IDLibro");
+                datos.SetearParametro("IDCarrito", idCarrito);
+                datos.SetearParametro("IDLibro", IdLibro);
+
+                int count = (int)datos.EjecutarScalar();
+                return count > 0;
+            }
+            finally
+            {
+                datos.Cerrar();
+            }
+        }
         public void Agregar(LibroPorCarrito nuevo)
         {
             datos = new AccesoDatos();
             try
             {
                 datos.Conectar();
-                datos.Consultar("INSERT INTO LibrosPorCarrito (IDCarrito, IDLibro, Cantidad, PrecioUnitario) " +
-                                "VALUES (@IDCarrito, @IDLibro, @Cantidad, @PrecioUnitario)");
-                datos.SetearParametro("@IDCarrito", nuevo.IdCarrito);
-                datos.SetearParametro("@IDLibro", nuevo.IdLibro);
-                datos.SetearParametro("@Cantidad", nuevo.Cantidad);
-                datos.SetearParametro("@PrecioUnitario", nuevo.PrecioUnitario);
-
+                if(ExisteElLibro(nuevo.IdCarrito,nuevo.IdLibro))
+                {
+                    datos.Consultar("UPDATE LibrosPorCarrito SET Cantidad = Cantidad + 1 WHERE IDCarrito = @IDCarrito AND IDLibro = @IDLibro");
+                    datos.SetearParametro("@IDCarrito", nuevo.IdCarrito);
+                    datos.SetearParametro("@IDLibro", nuevo.IdLibro);
+                }
+                else
+                {
+                    datos.Consultar("INSERT INTO LibrosPorCarrito (IDCarrito, IDLibro, Cantidad, PrecioUnitario) " +
+                                    "VALUES (@IDCarrito, @IDLibro, @Cantidad, @PrecioUnitario)");
+                    datos.SetearParametro("@IDCarrito", nuevo.IdCarrito);
+                    datos.SetearParametro("@IDLibro", nuevo.IdLibro);
+                    datos.SetearParametro("@Cantidad", nuevo.Cantidad);
+                    datos.SetearParametro("@PrecioUnitario", nuevo.PrecioUnitario);
+                }
+                
                 datos.EjecutarNonQuery();
             }
             catch (Exception ex)
@@ -69,10 +96,34 @@ namespace Negocio
             try
             {
                 datos.Conectar();
-                datos.Consultar("DELETE FROM LibrosPorCarrito WHERE IDCarrito = @IDCarrito AND IDLibro = @IDLibro");
-                datos.SetearParametro("@IDCarrito", idCarrito);
-                datos.SetearParametro("@IDLibro", idLibro);
-                datos.EjecutarNonQuery();
+
+                datos.Consultar("SELECT Cantidad FROM LibrosPorCarrito WHERE IDCarrito = @IDCarrito AND IDLibro = @IDLibro");
+                datos.SetearParametro("IDCarrito", idCarrito);
+                datos.SetearParametro("IDLibro", idLibro);
+
+                datos.Leer();
+
+                if(datos.Lector.Read())
+                {
+                    int cantidad = (int)datos.Lector["Cantidad"];
+                    datos.Cerrar();
+                    datos.Conectar();
+
+                    if (cantidad > 1)
+                    {
+                        datos.Consultar("UPDATE LibrosPorCarrito SET Cantidad = Cantidad - 1 WHERE IDCarrito = @IDCarrito AND IDLibro = @IDLibro");
+                        datos.SetearParametro("@IDCarrito", idCarrito);
+                        datos.SetearParametro("@IDLibro", idLibro);
+                        datos.EjecutarNonQuery();
+                    }
+                    else
+                    {
+                        datos.Consultar("DELETE FROM LibrosPorCarrito WHERE IDCarrito = @IDCarrito AND IDLibro = @IDLibro");
+                        datos.SetearParametro("@IDCarrito", idCarrito);
+                        datos.SetearParametro("@IDLibro", idLibro);
+                        datos.EjecutarNonQuery();
+                    }
+                }
             }
             catch (Exception er)
             {
